# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "User API",
    "version": "1.0.0",
    "description": "API documentation for user-related operations"
  },
  "servers": [
    {
      "url": "https://backend-production-7b98.up.railway.app"
    }
  ],
...
  "paths": {
    "/user/getUser/{userId}": {
      "get": {
        "tags": ["User"],
        "summary": "Retrieve a user by ID",
        "description": "Retrieve a user by their unique ID from the database.",
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique ID of the user"
          }
        ],
        "responses": {
          "200": {
            "description": "A user object",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          },
          "406": {
            "description": "Invalid user id"
          },
          "500": {
            "description": "Unable to find the user, try again"
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "tags": ["User"],
        "summary": "Verify credentials and send accessToken",
        "description": "Verify user credentials and send an access token.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "userId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Wrong password"
          },
          "404": {
            "description": "Cannot find user"
          },
          "500": {
            "description": "Unable to login, try again"
          }
        }
      }
    },
    "/user/register": {
      "post": {
        "tags": ["User"],
        "summary": "Register a user and encrypt the password",
        "description": "Register a new user and encrypt their password.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "userId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully"
          },
          "409": {
            "description": "Email or Username already exists"
          },
          "406": {
            "description": ["Invalid email or user id",  "The password must be 8 char long", "The password must contain a letter", "The password must contain a number", "The password must contain at least one special char between these: ? ! . _ - @ |"]
          },
          "500": {
            "description": "Unable to register the user, try again"
          }
        }
      }
    },
    "/user/modifyUsername": {
      "put": {
        "tags": ["User"],
        "summary": "Modify the username",
        "description": "Modify the username of an existing user.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string"
                  },
                  "newUsername": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated"
          },
          "409": {
            "description": "Username already exists"
          },
          "404": {
            "description": "Cannot find user"
          },
          "406": {
            "description": "Invalid user id"
          },
          "500": {
            "description": "Unable to update the username, try again"
          }
        }
      }
    },
    "/user/modifyPassword": {
      "put": {
        "tags": ["User"],
        "summary": "Modify the password",
        "description": "Modify the password of an existing user.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string"
                  },
                  "oldPassword": {
                    "type": "string"
                  },
                  "newPassword": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated"
          },
          "401": {
            "description": "Old password is incorrect"
          },
          "404": {
            "description": "Cannot find the user"
          },
          "406": {
            "description": ["Invalid user id", "The password must be 8 char long", "The password must contain a letter", "The password must contain a number", "The password must contain at least one special char between these: ? ! . _ - @ |"]
          },
          "500": {
            "description": "Unable to update the password, try again"
          }
        }
      }
    },
    "/user/deleteProfile/{userId}": {
      "delete": {
        "tags": ["User"],
        "summary": "Delete the user account",
        "description": "Delete the user account by their unique ID.",
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique ID of the user"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted"
          },
          "404": {
            "description": "Cannot find user"
          },
          "406": {
            "description": "Invalid user id"
          },
          "500": {
            "description": "Unable to delete the account, try again"
          }
        }
      }
    },
...
  }
}